# Enhanced nexmark cli and cleaned up code