--- a/benches/nexmark/main.rs
+++ b/benches/nexmark/main.rs
@@ -3,22 +3,17 @@
 //! CLI for running Nexmark benchmarks with DBSP.
 #![feature(is_some_with)]
 
-#[cfg(unix)]
-use libc::{getrusage, rusage, timeval};
-#[cfg(unix)]
-use std::{io::Error, mem::MaybeUninit};
-use std::{
-    sync::mpsc,
-    thread::{self, JoinHandle},
-    time::{Duration, Instant},
-};
+#[macro_use]
+mod run_queries;
+mod allocation;
 
+use crate::allocation::{AllocStats, MiMalloc};
 use anyhow::{anyhow, Result};
 use ascii_table::AsciiTable;
 use clap::Parser;
 use dbsp::{
     nexmark::{
-        config::Config as NexmarkConfig,
+        config::{Config as NexmarkConfig, Query as NexmarkQuery},
         model::Event,
         queries::{
             q0, q1, q12, q13, q13_side_input, q14, q15, q16, q17, q18, q19, q2, q20, q21, q22, q3,
@@ -31,44 +26,20 @@ use dbsp::{
 };
 use indicatif::{ProgressBar, ProgressStyle};
 use num_format::{Locale, ToFormattedString};
+use size_of::HumanBytes;
+use std::{
+    sync::mpsc,
+    thread::{self, JoinHandle},
+    time::{Duration, Instant},
+};
+
+#[global_allocator]
+static ALLOC: MiMalloc = MiMalloc;
 
 // TODO: Ideally these macros would be in a separate `lib.rs` in this benchmark
 // crate, but benchmark binaries don't appear to work like that (in that, I
 // haven't yet found a way to import from a `lib.rs` in the same directory as
 // the benchmark's `main.rs`)
-
-/// Returns a closure for a circuit with the nexmark source that returns
-/// the input handle.
-macro_rules! nexmark_circuit {
-    ( "q13", $q:expr ) => {
-        |circuit: &mut Circuit<()>| {
-            let (stream, input_handle) = circuit.add_input_zset::<Event, isize>();
-            let (side_stream, mut side_input_handle) =
-                circuit.add_input_zset::<(usize, String, u64), isize>();
-
-            let output = $q(stream, side_stream);
-
-            output.inspect(move |_zs| ());
-
-            // Ensure the side-input is loaded here so we can return a single input
-            // handle like the other queries.
-            side_input_handle.append(&mut q13_side_input());
-
-            input_handle
-        }
-    };
-    ( $q_name:expr, $q:expr ) => {
-        |circuit: &mut Circuit<()>| {
-            let (stream, input_handle) = circuit.add_input_zset::<Event, isize>();
-
-            let output = $q(stream);
-
-            output.inspect(move |_zs| ());
-
-            input_handle
-        }
-    };
-}
 
 /// Currently just the elapsed time, but later add CPU and Mem.
 #[derive(Default)]
@@ -76,9 +47,8 @@ struct NexmarkResult {
     name: String,
     num_events: u64,
     elapsed: Duration,
-    total_usr_cpu: Duration,
-    total_sys_cpu: Duration,
-    max_rss: Option<u64>,
+    before_stats: AllocStats,
+    after_stats: AllocStats,
 }
 
 struct InputStats {
@@ -206,100 +176,29 @@ fn coordinate_input_and_steps(
     }
 }
 
-macro_rules! run_query {
-    ( $q_name:tt, $q:expr, $nexmark_config:expr) => {{
-        // let circuit_closure = nexmark_circuit!($q_name, $q);
-        let circuit_closure = nexmark_circuit!($q_name, $q);
-
-        let num_cores = $nexmark_config.cpu_cores;
-        let expected_num_events = $nexmark_config.max_events;
-        let (dbsp, input_handle) = Runtime::init_circuit(num_cores, circuit_closure).unwrap();
-
-        // Create a channel for the coordinating thread to determine whether the
-        // producer or consumer step is completed first.
-        let (step_done_tx, step_done_rx) = mpsc::sync_channel(2);
-
-        // Start the DBSP runtime processing steps only when it receives a message to do
-        // so. The DBSP processing happens in its own thread where the resource usage
-        // calculation can also happen.
-        let (dbsp_step_tx, dbsp_step_rx) = mpsc::sync_channel(1);
-        let dbsp_join_handle = spawn_dbsp_consumer(dbsp, dbsp_step_rx, step_done_tx.clone());
-
-        // Start the generator inputting the specified number of batches to the circuit
-        // whenever it receives a message.
-        let (source_step_tx, source_step_rx): (mpsc::SyncSender<()>, mpsc::Receiver<()>) =
-            mpsc::sync_channel(1);
-        let (source_exhausted_tx, source_exhausted_rx) = mpsc::sync_channel(1);
-        spawn_source_producer(
-            $nexmark_config,
-            input_handle,
-            source_step_rx,
-            step_done_tx,
-            source_exhausted_tx,
-        );
-
-        let input_stats = coordinate_input_and_steps(
-            expected_num_events,
-            dbsp_step_tx,
-            source_step_tx,
-            step_done_rx,
-            source_exhausted_rx,
-            dbsp_join_handle,
-        )
-        .unwrap();
-
-        // Return the user/system CPU overhead from the generator/input thread.
-        NexmarkResult {
-            num_events: input_stats.num_events,
-            ..NexmarkResult::default()
-        }
-    }};
-}
-
-macro_rules! run_queries {
-    ( $nexmark_config:expr, $max_events:expr, $queries_to_run:expr, $( ($q_name:tt, $q:expr) ),+ ) => {{
-        let mut results: Vec<NexmarkResult> = Vec::new();
-        // We have no way (currently) of finding the max memory usage for each
-        // subsequent query as the value is for the process. So only the first
-        // query will have a value.
-        let mut query_count = 0;
-        $(
-        if $queries_to_run.len() == 0 || $queries_to_run.contains(&$q_name.to_string()) {
-            query_count += 1;
-            println!("Starting {} bench of {} events...", $q_name, $max_events);
-
-            let start = Instant::now();
-            let (before_usr_cpu, before_sys_cpu, before_max_rss) = unsafe { rusage(RUSAGE_SELF) };
-
-            let thread_nexmark_config = $nexmark_config.clone();
-            let result = run_query!($q_name, $q, thread_nexmark_config);
-            let (after_usr_cpu, after_sys_cpu, after_max_rss) = unsafe { rusage(RUSAGE_SELF) };
-            results.push(NexmarkResult {
-                name: $q_name.to_string(),
-                total_usr_cpu: after_usr_cpu - before_usr_cpu,
-                total_sys_cpu: after_sys_cpu - before_sys_cpu,
-                max_rss: match query_count { 1 => Some(after_max_rss - before_max_rss), _ => None},
-                elapsed: start.elapsed(),
-                ..result
-            });
-        }
-        )+
-        results
-    }};
-}
-
 fn create_ascii_table() -> AsciiTable {
     let mut ascii_table = AsciiTable::default();
     ascii_table.set_max_width(200);
-    ascii_table.column(0).set_header("Query");
-    ascii_table.column(1).set_header("#Events");
-    ascii_table.column(2).set_header("Cores");
-    ascii_table.column(3).set_header("Elapsed");
-    ascii_table.column(4).set_header("Cores * Elapsed");
-    ascii_table.column(5).set_header("Throughput/Cores");
-    ascii_table.column(6).set_header("Total Usr CPU");
-    ascii_table.column(7).set_header("Total Sys CPU");
-    ascii_table.column(8).set_header("Max RSS(Kb)");
+
+    let columns = [
+        "Query",
+        "#Events",
+        "Cores",
+        "Elapsed",
+        "Cores * Elapsed",
+        "Throughput/Cores",
+        "Total Usr CPU",
+        "Total Sys CPU",
+        "Current RSS",
+        "Peak RSS",
+        "Current Commit",
+        "Peak Commit",
+        "Page Faults",
+    ];
+    for (idx, column_name) in columns.into_iter().enumerate() {
+        ascii_table.column(idx).set_header(column_name);
+    }
+
     ascii_table
 }
 
@@ -323,90 +222,60 @@ fn main() -> Result<()> {
         nexmark_config,
         max_events,
         queries_to_run,
-        ("q0", q0),
-        ("q1", q1),
-        ("q2", q2),
-        ("q3", q3),
-        ("q4", q4),
-        ("q5", q5),
-        ("q6", q6),
-        ("q7", q7),
-        ("q8", q8),
-        ("q9", q9),
-        ("q12", q12),
-        ("q13", q13),
-        ("q14", q14),
-        ("q15", q15),
-        ("q16", q16),
-        ("q17", q17),
-        ("q18", q18),
-        ("q19", q19),
-        ("q20", q20),
-        ("q21", q21),
-        ("q22", q22)
+        queries => {
+            q0,
+            q1,
+            q2,
+            q3,
+            q4,
+            q5,
+            q6,
+            q7,
+            q8,
+            q9,
+            q12,
+            q13,
+            q14,
+            q15,
+            q16,
+            q17,
+            q18,
+            q19,
+            q20,
+            q21,
+            q22,
+        }
     );
 
     let ascii_table = create_ascii_table();
-    ascii_table.print(results.into_iter().map(|r| {
+    ascii_table.print(results.into_iter().map(|result| {
+        let (before, after) = (result.before_stats, result.after_stats);
+
         vec![
-            r.name,
-            format!("{}", r.num_events.to_formatted_string(&Locale::en)),
+            result.name,
+            format!("{}", result.num_events.to_formatted_string(&Locale::en)),
             format!("{cpu_cores}"),
-            format!("{0:.3}s", r.elapsed.as_secs_f32()),
-            format!("{0:.3}s", cpu_cores as f32 * r.elapsed.as_secs_f32()),
+            format!("{:#.3?}", result.elapsed),
+            format!("{:#.3?}", result.elapsed * cpu_cores as u32),
             format!(
                 "{0:.3} K/s",
-                r.num_events as f32 / r.elapsed.as_secs_f32() / cpu_cores as f32 / 1000.0
+                result.num_events as f32 / result.elapsed.as_secs_f32() / cpu_cores as f32 / 1000.0,
             ),
-            format!("{0:.3}s", (r.total_usr_cpu).as_secs_f32()),
-            format!("{0:.3}s", (r.total_sys_cpu).as_secs_f32()),
             format!(
-                "{}",
-                if let Some(max_rss) = r.max_rss {
-                    max_rss.to_formatted_string(&Locale::en)
-                } else {
-                    "N/A".to_string()
-                }
+                "{:#.3?}",
+                Duration::from_millis((after.user_ms - before.user_ms) as u64),
             ),
+            format!(
+                "{:#.3?}",
+                Duration::from_millis((after.system_ms - before.system_ms) as u64),
+            ),
+            format!("{}", HumanBytes::from(after.current_rss)),
+            format!("{}", HumanBytes::from(after.peak_rss)),
+            format!("{}", HumanBytes::from(after.current_commit)),
+            format!("{}", HumanBytes::from(after.peak_commit)),
+            format!("{}", after.page_faults - before.page_faults),
         ]
     }));
 
     Ok(())
-}
-
-#[cfg(unix)]
-const RUSAGE_SELF: i32 = libc::RUSAGE_SELF;
-
-#[cfg(not(unix))]
-const RUSAGE_SELF: i32 = 0;
-
-#[cfg(unix)]
-fn duration_for_timeval(tv: timeval) -> Duration {
-    Duration::new(tv.tv_sec as u64, tv.tv_usec as u32 * 1_000)
-}
-
-/// Returns the user CPU, system CPU and maxrss (in Kb) for the current process.
-#[cfg(unix)]
-pub unsafe fn rusage(target: i32) -> (Duration, Duration, u64) {
-    let mut ru: MaybeUninit<rusage> = MaybeUninit::uninit();
-    let err_code = getrusage(target, ru.as_mut_ptr());
-    if err_code != 0 {
-        panic!("getrusage returned {}", Error::last_os_error());
-    }
-    let ru = ru.assume_init();
-    (
-        duration_for_timeval(ru.ru_utime),
-        duration_for_timeval(ru.ru_stime),
-        ru.ru_maxrss as u64,
-    )
-}
-
-/// Define for non-unix so that bench can still run reporting on time etc.
-///
-/// # Safety
-///
-/// N/A
-#[cfg(not(unix))]
-pub unsafe fn rusage(_target: i32) -> (Duration, Duration, u64) {
-    (Duration::from_millis(0), Duration::from_millis(0), 0)
 }