This is an extraction of code from #193 since I got kinda sidetracked there

* Switched nexmark to use [mimalloc](https://github.com/microsoft/mimalloc) for the global allocator
* Cleaned up nexmark benchmark code
* Enhanced nexmark cli

I enhanced nexmark query parsing somewhat, where previously invalid query names passed to `--query` would result in a silent failure (that is, using `--query something_invalid` would simply do nothing), it now knows what the available queries are and therefore can validate them, which leads to this behavior now
```
λ  cargo bench --bench nexmark ... --query something_invalid
error: "something_invalid" isn't a valid value for '--query <QUERY>'
        [possible values: q0, q1, q2, q3, q4, q5, q6, q7, q8, q9, q12, q13, q14, q15, q16, q17, q18, q19, q20, q21, q22]

For more information try --help
error: bench failed, to rerun pass `--bench nexmark`
```
Additionally the help prompt now tells you the available queries
```
λ  cargo bench --bench nexmark --help
...
        --query <QUERY>
            Queries to run, all by default

            [env: NEXMARK_QUERIES=]
            [possible values: q0, q1, q2, q3, q4, q5, q6, q7, q8, q9, q12, q13, q14, q15, q16, q17,
            q18, q19, q20, q21, q22]
...
```
I also took advantage of using `mimalloc` to get more/richer/better stats on each query
```
┌───────┬─────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┬────────────┬────────────────┬─────────────┬─────────────┐
│ Query │ #Events     │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Current RSS │ Peak RSS   │ Current Commit │ Peak Commit │ Page Faults │
├───────┼─────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┼────────────┼────────────────┼─────────────┼─────────────┤
│ q22   │ 100,000,000 │ 8     │ 14.426s │ 115.412s        │ 866.462 K/s      │ 129.468s      │ 203.000ms     │ 140.55 MiB  │ 142.25 MiB │ 401.77 MiB     │ 408.84 MiB  │ 57885       │
└───────┴─────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┴────────────┴────────────────┴─────────────┴─────────────┘
```