@ryzhyk 2022-09-28
This yields an impressive speedup on my Linux laptop (Core i7-12700H, 64GB RAM).

`main` branch:

```
cargo bench --bench nexmark --features with-nexmark -- --first-event-rate=10000000 --max-events=100000000 --cpu-cores 8  --num-event-generators 6 --source-buffer-size 10000 --input-batch-size 40000
    Finished bench [optimized + debuginfo] target(s) in 0.07s
     Running benches/nexmark/main.rs (target/release/deps/nexmark-5744354f98061c26)
Starting q0 bench of 100000000 events...
┌───────┬─────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┐
│ Query │ #Events     │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Max RSS(Kb) │
├───────┼─────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┤
│ q0    │ 100,000,000 │ 8     │ 13.360s │ 106.878s        │ 935.645 K/s      │ 127.636s      │ 4.148s        │ 67,964      │
│ q1    │ 100,000,000 │ 8     │ 14.601s │ 116.811s        │ 856.080 K/s      │ 148.708s      │ 4.290s        │ N/A         │
│ q2    │ 100,000,000 │ 8     │ 14.308s │ 114.463s        │ 873.642 K/s      │ 139.634s      │ 4.424s        │ N/A         │
│ q3    │ 100,000,000 │ 8     │ 15.557s │ 124.454s        │ 803.511 K/s      │ 144.117s      │ 5.210s        │ N/A         │
│ q4    │ 100,000,000 │ 8     │ 19.423s │ 155.385s        │ 643.564 K/s      │ 179.126s      │ 9.382s        │ N/A         │
│ q5    │ 100,000,000 │ 8     │ 15.262s │ 122.095s        │ 819.034 K/s      │ 147.878s      │ 4.889s        │ N/A         │
│ q6    │ 100,000,000 │ 8     │ 20.272s │ 162.178s        │ 616.606 K/s      │ 188.824s      │ 9.010s        │ N/A         │
│ q7    │ 100,000,000 │ 8     │ 23.810s │ 190.482s        │ 524.985 K/s      │ 186.988s      │ 12.588s       │ N/A         │
│ q8    │ 100,000,000 │ 8     │ 16.687s │ 133.494s        │ 749.100 K/s      │ 145.608s      │ 5.021s        │ N/A         │
│ q9    │ 100,000,000 │ 8     │ 76.459s │ 611.669s        │ 163.487 K/s      │ 307.454s      │ 14.660s       │ N/A         │
│ q12   │ 100,000,000 │ 8     │ 16.516s │ 132.128s        │ 756.842 K/s      │ 148.527s      │ 4.854s        │ N/A         │
│ q13   │ 100,000,000 │ 8     │ 21.198s │ 169.587s        │ 589.667 K/s      │ 183.184s      │ 10.860s       │ N/A         │
│ q14   │ 100,000,000 │ 8     │ 14.777s │ 118.215s        │ 845.919 K/s      │ 154.526s      │ 4.306s        │ N/A         │
│ q15   │ 100,000,000 │ 8     │ 17.959s │ 143.674s        │ 696.019 K/s      │ 200.183s      │ 7.489s        │ N/A         │
│ q16   │ 100,000,000 │ 8     │ 56.560s │ 452.483s        │ 221.003 K/s      │ 422.147s      │ 8.820s        │ N/A         │
│ q17   │ 100,000,000 │ 8     │ 26.083s │ 208.661s        │ 479.247 K/s      │ 234.232s      │ 12.051s       │ N/A         │
│ q18   │ 100,000,000 │ 8     │ 41.815s │ 334.522s        │ 298.934 K/s      │ 222.287s      │ 13.447s       │ N/A         │
│ q19   │ 100,000,000 │ 8     │ 43.234s │ 345.875s        │ 289.122 K/s      │ 283.740s      │ 19.375s       │ N/A         │
│ q20   │ 100,000,000 │ 8     │ 32.521s │ 260.169s        │ 384.365 K/s      │ 212.776s      │ 11.571s       │ N/A         │
│ q21   │ 100,000,000 │ 8     │ 16.865s │ 134.920s        │ 741.182 K/s      │ 155.587s      │ 4.916s        │ N/A         │
│ q22   │ 100,000,000 │ 8     │ 16.670s │ 133.357s        │ 749.869 K/s      │ 169.004s      │ 12.899s       │ N/A         │
└───────┴─────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┘
```

With this PR:

```
┌───────┬─────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┬────────────┬────────────────┬─────────────┬─────────────┐
│ Query │ #Events     │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Current RSS │ Peak RSS   │ Current Commit │ Peak Commit │ Page Faults │
├───────┼─────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┼────────────┼────────────────┼─────────────┼─────────────┤
│ q0    │ 100,000,000 │ 8     │ 10.658s │ 85.268s         │ 1172.778 K/s     │ 98.951s       │ 589.000ms     │ 474.44 MiB  │ 162.62 MiB │ 474.44 MiB     │ 474.47 MiB  │ 0           │
│ q1    │ 100,000,000 │ 8     │ 11.436s │ 91.488s         │ 1093.038 K/s     │ 114.388s      │ 802.000ms     │ 483.19 MiB  │ 215.74 MiB │ 483.19 MiB     │ 483.22 MiB  │ 0           │
│ q2    │ 100,000,000 │ 8     │ 11.422s │ 91.380s         │ 1094.335 K/s     │ 106.085s      │ 663.000ms     │ 493.12 MiB  │ 238.32 MiB │ 493.12 MiB     │ 493.16 MiB  │ 0           │
│ q3    │ 100,000,000 │ 8     │ 12.402s │ 99.214s         │ 1007.924 K/s     │ 116.652s      │ 951.000ms     │ 766.50 MiB  │ 483.72 MiB │ 766.50 MiB     │ 766.54 MiB  │ 4           │
│ q4    │ 100,000,000 │ 8     │ 15.402s │ 123.216s        │ 811.581 K/s      │ 152.873s      │ 4.450s        │ 5.27 GiB    │ 7.74 GiB   │ 5.27 GiB       │ 8.11 GiB    │ 0           │
│ q5    │ 100,000,000 │ 8     │ 12.373s │ 98.985s         │ 1010.253 K/s     │ 116.545s      │ 1.100s        │ 5.28 GiB    │ 7.74 GiB   │ 5.28 GiB       │ 8.11 GiB    │ 4           │
│ q6    │ 100,000,000 │ 8     │ 15.705s │ 125.642s        │ 795.911 K/s      │ 160.860s      │ 3.118s        │ 5.80 GiB    │ 8.31 GiB   │ 5.80 GiB       │ 8.62 GiB    │ 0           │
│ q7    │ 100,000,000 │ 8     │ 17.235s │ 137.877s        │ 725.282 K/s      │ 171.736s      │ 5.004s        │ 10.06 GiB   │ 15.14 GiB  │ 10.06 GiB      │ 15.88 GiB   │ 0           │
│ q8    │ 100,000,000 │ 8     │ 12.894s │ 103.150s        │ 969.460 K/s      │ 117.957s      │ 1.144s        │ 10.08 GiB   │ 15.14 GiB  │ 10.08 GiB      │ 15.88 GiB   │ 0           │
│ q9    │ 100,000,000 │ 8     │ 63.720s │ 509.761s        │ 196.170 K/s      │ 279.577s      │ 7.225s        │ 16.60 GiB   │ 22.29 GiB  │ 16.60 GiB      │ 22.98 GiB   │ 0           │
│ q12   │ 100,000,000 │ 8     │ 13.090s │ 104.717s        │ 954.952 K/s      │ 125.338s      │ 832.000ms     │ 16.61 GiB   │ 22.29 GiB  │ 16.61 GiB      │ 22.98 GiB   │ 0           │
│ q13   │ 100,000,000 │ 8     │ 19.497s │ 155.975s        │ 641.128 K/s      │ 177.251s      │ 4.519s        │ 16.63 GiB   │ 22.29 GiB  │ 16.63 GiB      │ 22.98 GiB   │ 0           │
│ q14   │ 100,000,000 │ 8     │ 12.498s │ 99.984s         │ 1000.159 K/s     │ 122.887s      │ 741.000ms     │ 16.64 GiB   │ 22.29 GiB  │ 16.64 GiB      │ 22.98 GiB   │ 0           │
│ q15   │ 100,000,000 │ 8     │ 14.248s │ 113.987s        │ 877.291 K/s      │ 170.583s      │ 2.247s        │ 16.65 GiB   │ 22.29 GiB  │ 16.65 GiB      │ 22.98 GiB   │ 7           │
│ q16   │ 100,000,000 │ 8     │ 50.588s │ 404.705s        │ 247.094 K/s      │ 397.774s      │ 4.699s        │ 16.67 GiB   │ 22.29 GiB  │ 16.67 GiB      │ 22.98 GiB   │ 0           │
│ q17   │ 100,000,000 │ 8     │ 19.736s │ 157.888s        │ 633.359 K/s      │ 224.041s      │ 2.370s        │ 17.92 GiB   │ 22.29 GiB  │ 17.92 GiB      │ 22.98 GiB   │ 0           │
│ q18   │ 100,000,000 │ 8     │ 37.111s │ 296.889s        │ 336.826 K/s      │ 205.415s      │ 5.691s        │ 17.92 GiB   │ 23.98 GiB  │ 17.92 GiB      │ 24.38 GiB   │ 0           │
│ q19   │ 100,000,000 │ 8     │ 34.253s │ 274.027s        │ 364.928 K/s      │ 262.645s      │ 7.180s        │ 17.92 GiB   │ 25.42 GiB  │ 17.92 GiB      │ 26.01 GiB   │ 0           │
│ q20   │ 100,000,000 │ 8     │ 26.099s │ 208.789s        │ 478.953 K/s      │ 198.890s      │ 5.835s        │ 17.92 GiB   │ 25.42 GiB  │ 17.92 GiB      │ 26.01 GiB   │ 0           │
│ q21   │ 100,000,000 │ 8     │ 12.916s │ 103.327s        │ 967.804 K/s      │ 122.273s      │ 825.000ms     │ 17.92 GiB   │ 25.42 GiB  │ 17.92 GiB      │ 26.01 GiB   │ 0           │
│ q22   │ 100,000,000 │ 8     │ 12.324s │ 98.593s         │ 1014.268 K/s     │ 125.098s      │ 821.000ms     │ 17.92 GiB   │ 25.42 GiB  │ 17.92 GiB      │ 26.01 GiB   │ 0           │
└───────┴─────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┴────────────┴────────────────┴─────────────┴─────────────┘

```